+++ src/apis/artwork/getArtworkList.ts
@@ -1,10 +1,7 @@
-import { useQuery } from '@tanstack/react-query';
-
 import defaultClient from '@/apis';
 import { ARTWORK } from '@/constants/API';
-import { ArtworkListResponse, PaginationType } from '@/types';
 
-interface ArtworkListParams {
+export interface ArtworkListParams {
   sort: string;
   artworkField?: string;
   search: string;
@@ -43,18 +40,4 @@ const getArtworkList = async ({
   }
 };
 
-const useArtworkList = (params: ArtworkListParams) => {
-  const { data, isLoading, error } = useQuery<ArtworkListResponse>({
-    queryKey: [ARTWORK.artworkList, params],
-    queryFn: () => getArtworkList({ ...params }),
-    retry: 3,
-  });
-
-  return {
-    data: data || { data: [], page: {} as PaginationType },
-    isLoading,
-    error,
-  };
-};
-
-export default useArtworkList;
+export default getArtworkList;
