+++ src/app/artwork/list/page.tsx
@@ -2,11 +2,11 @@
 
 import { useState } from 'react';
 
-import useArtworkList from '@/apis/artwork/getArtworkList';
 import ArtworkFilter from '@/components/artwork/ArtworkFilter';
 import ArtworkSearchBar from '@/components/artwork/ArtworkSearchBar';
 import ArtworkShowcase from '@/components/artwork/ArtworkShowcase';
 import FollowedArtistsSection from '@/components/artwork/FollowedArtistsSection';
+import useArtworkList from '@/hooks/serverStateHooks/useArtworkList';
 
 const SORT_MAPPING: Record<string, string> = {
   최신순: 'recent',
